CODE/FRONTEND/requirements.txt
Flask
flask-cors
numpy
pandas
scikit-learn
imbalanced-learn
python-dotenv
google-generativeai
requests
gunicorn
